tests/fuzzers: fixup
