+++ tests/fuzzers/whisperv6/whisper-fuzzer.go
@@ -1,3 +1,19 @@
+// Copyright 2019 The go-ethereum Authors
+// This file is part of the go-ethereum library.
+//
+// The go-ethereum library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The go-ethereum library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
+
 package whisperv6
 
 import (
@@ -60,7 +76,6 @@ func Fuzz(input []byte) int {
 	}
 	if !decrypted.ValidateAndParse() {
 		panic("ValidateAndParse failed")
-		return 0
 	}
 	if !bytes.Equal(text, decrypted.Payload) {
 		panic("text != decrypted.Payload")
@@ -71,6 +86,5 @@ func Fuzz(input []byte) int {
 	if !whisperv6.IsPubKeyEqual(decrypted.Src, &params.Src.PublicKey) {
 		panic("Unexpected public key")
 	}
-
 	return 0
 }
